requirements.txt
django==2.2.28
mysqlclient==2.1.1
pillow==10.2.0
uwsgi
